[ntuple] Move relevant implementations to new RFieldSequenceContainer.cxx
Implementations of classes in RFieldSequenceContainer.hxx introduced
by commit 8242ab4c79.
