+++ tree/ntuple/CMakeLists.txt
@@ -60,6 +60,7 @@ SOURCES
   v7/src/RField.cxx
   v7/src/RFieldBase.cxx
   v7/src/RFieldMeta.cxx
+  v7/src/RFieldSequenceContainer.cxx
   v7/src/RFieldUtils.cxx
   v7/src/RFieldVisitor.cxx
   v7/src/RMiniFile.cxx
